fix: combine scripts and try replace in chunk*.js
