+++ packages/components/package.json
@@ -51,9 +51,8 @@
     "build.package": "pnpm build.package/create-index && vite build --config vite.config.package.js && pnpm build.package/remove-index",
     "build.package/create-index": "node scripts/node-generate-package-index.cjs && echo '✅ Index created'",
     "build.package/remove-index": "node -e \"require('fs').unlinkSync('src/solid-components.package.ts')\" && echo '✅ Index removed'",
-    "build.storybook": "storybook build --output-dir dist/storybook && cp dist/custom-elements.json dist/storybook/custom-elements.json && pnpm build.storybook-page-title && pnpm build.storybook-favicon",
-    "build.storybook-page-title": "node scripts/storybook/node-modify-page-title.cjs",
-    "build.storybook-favicon": "node scripts/storybook/node-modify-favicon.cjs",
+    "build.storybook": "storybook build --output-dir dist/storybook && cp dist/custom-elements.json dist/storybook/custom-elements.json && pnpm build.storybook/modify-storybook",
+    "build.storybook/modify-storybook": "node scripts/storybook/node-modify-storybook.cjs",
     "build.chromatic": "pnpm build.storybook && pnpm ci.optimize-chromatic",
     "preview": "vite preview",
     "verify": "pnpm ts.verify && pnpm build && echo '✅ Build verified' && pnpm ci.optimize-chromatic && echo '✅ Chromatic optimization verified' && pnpm test.verify && echo '✅ Test verified'",
